issue fixes
algorithm change for
page bounds calculation in case of bounded fling.

jump to implementation changed from
running in (n) to (1)
for both horizontal and vertical

added scroll bar movement for both horizontal / vertical
(this was done previously but at that time we were drawing all the pages.
now pages are selected based on screen size)

implementation of pages drawn based on screen size.