issue fix related to scaling
as we load some two to three pages in both directions top and bottom in order to avoid junk and we don't draw the content. (just page backgrounds)

so when scale happens it is delivered to the visible pages and if the scale and scroll event is combined.
the page snapshot is not generated for the newly visible page if the user is scrolling and scaling.

#Fix that by drawing the page content.